fix bugs

+++ database/migrations/2021_02_21_060618_create_detail_penjualans_table.php
@@ -20,7 +20,7 @@ class CreateDetailPenjualansTable extends Migration
             $table->integer('quantity');
             $table->decimal('harga_jual', 15, 2);
             $table->tinyInteger('diskon');
-            $table->double('subtotal_harga');
+            $table->decimal('subtotal_harga', 20, 2);
             $table->timestamps();
         });
     }
